FIX: com_remote.cxx deserializes int64 correctly
Not relevant for Java based API jobs.
Used for Object_id (the first 2**31 proxies are not affected)

+++ kernel-cpp/zschimmer/com_remote.cxx
@@ -2796,7 +2796,7 @@ int32 Input_message::read_int32()
 int64 Input_message::read_int64()
 {
     int32 first = read_int32();
-    return ( int64(first) << 32 ) | read_int32();
+    return (int64(first) << 32) | (uint32)read_int32();
 }
 
 //-----------------------------------------------------------------------Input_message::read_double
